# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: PW Backend API - Gestione Gioielleria
  description: API per gestione inventario articoli, ordini, fornitori e analytics
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Development Server
...
paths:
...
  /api/articoli:
    get:
      tags:
        - Articoli
      summary: Lista tutti gli articoli
      description: Recupera l'elenco completo degli articoli con filtri opzionali
      operationId: getArticoli
      parameters:
        - name: gruppo
          in: query
          required: false
          schema:
            type: string
        - name: stato
          in: query
          required: false
          schema:
            type: string
        - name: stagione
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Lista articoli
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Articolo'

    post:
      tags:
        - Articoli
      summary: Crea nuovo articolo
      description: Aggiunge un nuovo articolo all'inventario
      operationId: createArticolo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticoloRequest'
      responses:
        '201':
          description: Articolo creato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Articolo'
...
  /api/articoli/search:
    post:
      tags:
        - Articoli
      summary: Ricerca articoli
      description: Cerca articoli tramite searchbar e filtri
      operationId: searchArticoli
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Risultati ricerca
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Articolo'
...
components:
  schemas:
...
    Articolo:
      type: object
      properties:
        codice:
          type: string
          example: "ANL001"
        descrizione:
          type: string
          example: "Anello Solitario Diamante 0.50ct"
        ean:
          type: string
          example: "8001234567890"
        prezzodilistino:
          type: number
          format: double
          example: 2599.00
        gruppo:
          type: string
          example: "ANELLI"
        stato:
          type: string
          example: "AT"
        lp1:
          type: string
          example: "GIO"
        lp2:
          type: string
          example: "ANL"
        lp3:
          type: string
          example: "DIA"
        lp4:
          type: string
        lp5:
          type: string
        lineaprod:
          type: string
          example: "Anelli Diamanti"
        stagione:
          type: string
          example: "2026 S/S"
        linkimmagine:
          type: string
        bidone:
          type: string
        scodescri:
          type: string
          example: "Anello Solitr"
        tipo:
          type: string
          example: "ST"
        iva:
          type: string
          example: "22"
        codfornitore:
          type: string
        peso:
          type: string
        note:
          type: string
        ubicazione:
          type: string
          example: "SCAFF-A12"
        datainserimento:
          type: string
          format: date-time
        dataritiro:
          type: string
          format: date-time
        scortaminima:
          type: integer
        codpadre:
          type: string
        qtafiglio:
          type: integer
        codaccessori:
          type: string
        grcassa:
          type: string
        ordinabile:
          type: string
        datascad:
          type: string
          format: date-time
        cod_iniziale:
          type: string
        var1:
          type: string
        var2:
          type: string
        var3:
          type: string
        var4:
          type: string
        cursoremod:
          type: string
        gestgiacenza:
          type: string
        codice2:
          type: string
        confezione:
          type: string
        g_cliente:
          type: string
        progre:
          type: integer
        dataultimamodifica:
          type: string
          format: date-time
        online_relevant:
          type: integer
    ArticoloRequest:
      type: object
      required:
        - codice
        - descrizione
        - prezzodilistino
        - gruppo
        - stato
      properties:
        codice:
          type: string
        descrizione:
          type: string
        ean:
          type: string
        prezzodilistino:
          type: number
          format: double
        gruppo:
          type: string
        stato:
          type: string
        lp1:
          type: string
        lp2:
          type: string
        lp3:
          type: string
        lp4:
          type: string
        lp5:
          type: string
        lineaprod:
          type: string
        stagione:
          type: string
        linkimmagine:
          type: string
        scodescri:
          type: string
        tipo:
          type: string
        iva:
          type: string
        codfornitore:
          type: string
        peso:
          type: string
        note:
          type: string
        ubicazione:
          type: string
        scortaminima:
          type: integer
        ordinabile:
          type: string
        online_relevant:
          type: integer
    SearchRequest:
      type: object
      properties:
        query:
          type: string
          example: "anello"
        gruppo:
          type: string
          example: "ANELLI"
        prezzoMin:
          type: number
          format: double
        prezzoMax:
          type: number
          format: double
        stato:
          type: string
        stagione:
          type: string
...